Refactor to use terraform modules
