+++ Makefile
@@ -1,10 +1,4 @@
-build: build-rotate build-cleanup build-test
-
-plan:
-	cd terraform && terraform plan
-
-deploy:
-	cd terraform && terraform apply
+build: build-rotate build-cleanup
 
 build-rotate:
 	GOOS=linux go build -o build/rotate-lambda lambda/rotator/*.go
@@ -12,8 +6,5 @@ build-rotate:
 build-cleanup:
 	GOOS=linux go build -o build/cleanup-lambda lambda/cleanup/*.go
 
-build-test:
-	GOOS=linux go build -o build/tester-lambda lambda/tester/*.go
-
 clean:
-	rm build/*
+	rm build/*.zip
