+++ rotator/main.tf
@@ -1,0 +1,1 @@
+data "aws_caller_identity" "current" {}
