+++ rotator/outputs.tf
@@ -1,0 +1,3 @@
+output "rotator_lambda_arn" {
+  value = aws_lambda_function.rotate_lambda.arn
+}
