--- terraform/test
@@ -1,1 +1,0 @@
-it works!
